% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colour_thesaurus}
\alias{colour_thesaurus}
\title{Colour thesaurus}
\format{
A data frame with 240 rows and 2 variables:
\describe{
\item{colour}{preferred colour name: \code{"white"}, \code{"tan"}, \code{"yellow"},
\code{"orange"}, \code{"red"}, \code{"pink"}, \code{"purple"}, \code{"blue"}, \code{"green"}, \code{"brown"},
\code{"grey"}, or \code{"black"}}
\item{shade}{variant names for shades of \code{colour}}
}
}
\source{
\url{https://ingridsundberg.com/2014/02/04/the-color-thesaurus/}
}
\usage{
colour_thesaurus
}
\description{
A dataset of colour names compiled by Ingrid Sundberg, formatted as an
example thesaurus for use in \code{\link[=control]{control()}}. Contains 240 names for shades of
12 colours.
}
\keyword{datasets}
