% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controller-package.R
\docType{package}
\name{controller-package}
\alias{controller}
\alias{controller-package}
\title{controller: Controlled Vocabularies}
\description{
Provides functions for working with controlled vocabularies. A controlled vocabulary is a predefined set of terms intended to be used in in a specific context. This package introduces the 'control' verb, which recodes values in a vector using a lookup table of preferred and variant terms (a 'thesaurus'). It also includes functions for reading controlled vocabularies into R from standard formats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://controller.joeroe.io}
  \item \url{https://github.com/joeroe/controller}
  \item Report bugs at \url{https://github.com/joeroe/controller/issues}
}

}
\author{
\strong{Maintainer}: Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID}) [copyright holder]
}

}
\keyword{internal}
