% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controller-package.R
\docType{package}
\name{controller-package}
\alias{controller}
\alias{controller-package}
\title{controller: Controlled Vocabularies and Authority Control}
\description{
Functions for working with controlled vocabularies and authority
    control.
}
\author{
\strong{Maintainer}: Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID})

}
\keyword{internal}
