% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{read_fish}
\alias{read_fish}
\title{Read a FISH controlled vocabulary}
\usage{
read_fish(path)
}
\arguments{
\item{path}{Path or URL to a vocabulary in FISH's CSV format. Can be
either a \code{.zip} archive or a directory containing already uncompressed files.}
}
\value{
A data frame with two columns: \code{preferred} and \code{term}.
}
\description{
Reads controlled vocabularies from Historic England's FISH (Forum on
Information Standards in Heritage) group. These can be downloaded from:
\url{http://www.heritage-standards.org.uk/fish-vocabularies/}.
}
\references{
\itemize{
\item Forum on Information Standards in Heritage, "FISH Thesaurus Table Structure",
Available from: \url{http://www.heritage-standards.org.uk/fish-vocabularies/}.
}
}
